+++ js-functions/move-ball-functions/animateBallDrop.js
